refactor: split queue.rs and move triton_vm
These are some changes to prepare for moving job_queue into its own
crate, once we have a neptune-core workspace.

src/job_queue/triton_vm is moved to src/triton_vm_job_queue.  It has
neptune-core specific logic, and thus belongs in neptune-core crate.

Changes:
+ break job-queue into separate files for better organization
+ move src/job_queue/triton_vm to src/triton_vm_job_queue.
+ remove the LogWhenDropped newtype
+ refactor queue.rs and comment thoroughly.
+ add doc-comment with examples for JobHandle::cancel()
