+++ src/job_queue/job_id.rs
@@ -1,0 +1,18 @@
+/// a randomly generated Job identifier
+#[derive(Debug, Clone, Copy)]
+pub struct JobId([u8; 12]);
+
+impl std::fmt::Display for JobId {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        for byte in &self.0 {
+            write!(f, "{:02x}", byte)?;
+        }
+        Ok(())
+    }
+}
+
+impl JobId {
+    pub(super) fn random() -> Self {
+        Self(rand::random())
+    }
+}
