--- src/job_queue/triton_vm/mod.rs
@@ -1,5 +1,0 @@
-mod triton_vm_job_queue;
-
-pub use triton_vm_job_queue::vm_job_queue;
-pub use triton_vm_job_queue::TritonVmJobPriority;
-pub use triton_vm_job_queue::TritonVmJobQueue;
