+++ src/lib.rs
@@ -37,6 +37,7 @@ pub mod peer_loop;
 pub mod prelude;
 pub mod rpc_auth;
 pub mod rpc_server;
+pub mod triton_vm_job_queue;
 pub mod util_types;
 
 #[cfg(test)]
