+++ src/models/blockchain/transaction/validity/tasm/claims/generate_lock_script_claim_template.rs
@@ -88,11 +88,11 @@ mod tests {
     use tasm_lib::twenty_first::bfe;
 
     use super::GenerateLockScriptClaimTemplate;
-    use crate::job_queue::triton_vm::TritonVmJobPriority;
-    use crate::job_queue::triton_vm::TritonVmJobQueue;
     use crate::models::blockchain::transaction::primitive_witness::PrimitiveWitness;
     use crate::models::blockchain::transaction::validity::proof_collection::ProofCollection;
     use crate::models::blockchain::transaction::validity::tasm::claims::new_claim::NewClaim;
+    use crate::triton_vm_job_queue::TritonVmJobPriority;
+    use crate::triton_vm_job_queue::TritonVmJobQueue;
 
     impl Function for GenerateLockScriptClaimTemplate {
         fn rust_shadow(
