+++ src/models/proof_abstractions/tasm/prover_job.rs
@@ -17,10 +17,10 @@ use tasm_lib::triton_vm::error::InstructionError;
 use tokio::io::AsyncWriteExt;
 
 use crate::config_models::network::Network;
+use crate::job_queue::channels::JobCancelReceiver;
 use crate::job_queue::traits::Job;
-use crate::job_queue::traits::JobCancelReceiver;
-use crate::job_queue::traits::JobCompletion;
 use crate::job_queue::traits::JobResult;
+use crate::job_queue::JobCompletion;
 use crate::macros::fn_name;
 use crate::macros::log_scope_duration;
 use crate::models::blockchain::transaction::transaction_proof::TransactionProofType;
